# src/api/openapi.yaml
openapi: 3.1.0
info:
  title: SAVED4 OpenAPI Spec
  description: API specification for SAVED4
  version: 0.1.0
  x-fiddle-import-file: true
paths:
...
  /vehicles/{vehicle_id}:
    get:
      summary: Get vehicle
      operationId: getVehicle
      description: Returns a vehicle identified by vehicle id.
      tags:
        - vehicles
      security: []
      parameters:
        - $ref: "#/components/parameters/vehicle_id"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
    delete:
      summary: Delete vehicle
      operationId: deleteVehicle
      description: Deletes a vehicle.
      tags:
        - vehicles
      security: []
      parameters:
        - $ref: "#/components/parameters/vehicle_id"
      responses:
        "204":
          description: The request was successful, but there is no content to return in
            the response.
          content:
            application/json:
              schema:
                type: "null"
                title: DeleteVehicleNoContent
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
    patch:
      summary: Update vehicle
      operationId: updateVehicle
      description: Updates a vehicle.
      tags:
        - vehicles
      security: []
      parameters:
        - $ref: "#/components/parameters/vehicle_id"
      responses:
        "200":
          description: The request was successful, and the server has returned the
            requested resource in the response body.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerErrror"
      requestBody:
        required: true
        description: JSON object containing the fields to update
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehiclePartialUpdate"
...
components:
  parameters:
...
    vehicle_id:
      name: vehicle_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: The server could not understand the request due to invalid syntax.
        The client should modify the request and try again.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: array
                items:
                  type: object
                  properties:
                    message:
                      type: string
                  required:
                    - message
            required:
              - message
...
    InternalServerErrror:
      description: The server encountered an unexpected condition that prevented it
        from fulfilling the request. Report the issue to the support team if it
        persists.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
    NotFound:
      description: The server cannot find the requested resource. The endpoint may be
        invalid or the resource may no longer exist.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
            required:
              - message
...
  securitySchemes:
    BearerJwt:
      type: http
      scheme: Bearer
      bearerFormat: Json Web Token (JWT)
      description: "Jwt Auth: Authenticated requests contain a valid Json Web Token
        (JWT) as part of the `Authorization: Bearer <JWT>` header."
  schemas:
...
    Vehicle:
      type: object
      properties:
        vehicle_id:
          type: integer
          description: Unique identifier for the vehicle.
          example: 42
        position:
          type: object
          description: Current GPS location of the vehicle.
          properties:
            latitude:
              type: number
              description: Latitude of the vehicle's current position.
              example: 51.9244
            longitude:
              type: number
              description: Longitude of the vehicle's current position.
              example: 4.4777
          required:
            - latitude
            - longitude
        battery_level:
          type: integer
          description: Battery percentage of the electric vehicle.
          example: 76
        status:
          type: string
          enum:
            - available
            - in_use
            - out_of_order
          description: Current operational status (e.g., idle, en_route, charging).
          example: available
        order_id:
          type: integer
          description: Order currently assigned to the vehicle, if any.
          example: 203
      required:
        - vehicle_id
        - position
        - battery_level
        - status
      description: An EV that transports containers
    VehiclePartialUpdate:
      type: object
      properties:
        position:
          type: object
          description: Updated GPS location of the vehicle.
          properties:
            latitude:
              type: number
              description: Updated latitude of the vehicle.
              example: 52.3676
            longitude:
              type: number
              description: Updated longitude of the vehicle.
              example: 4.9041
          required:
            - latitude
            - longitude
        battery_level:
          type: integer
          description: Updated battery percentage of the vehicle.
          example: 65
        status:
          type: string
          enum:
            - available
            - in_use
            - out_of_order
          description: Updated operational status.
          example: in_use
        order_id:
          type: integer
          description: Updated ID of the order assigned to the vehicle.
          example: 205
      required: []
      description: Data transfer object for partially updating an existing Vehicle
        (PATCH operation).
      x-fiddle-dto-info:
        baseSchemaName: Vehicle
...